update_binary-addons: add -x option to skip no tag updates
without using the -x option, all binary addons are updated to the
latest hash is the branch if there are no tags set.

using the -x option, the binary addons are only updated to a newer
tag. when only a hash is present do not update the PKG_VERSION.

existing default behaviour of the script has not changed.

Co-authored-by: Matthias Reichl <github@hias.horus.com>